sdks/python/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]

[project]
name = "usm-secrets"
description = "Python SDK for Universal Secrets Manager"
readme = "README.md"
authors = [{name = "USM Contributors"}]
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dynamic = ["version"]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/universal-secrets-manager/usm"
Repository = "https://github.com/universal-secrets-manager/usm.git"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"